// src/drivers/dto/driver-input.dto.ts
import { VehicleFeature } from "../types/driver.type";

export type DriverInputDtoType = {
  name: string;
  phoneNumber: string;
  email: string;
  vehicleMake: string;
  vehicleModel: string;
  vehicleYear: number;
  vehicleLicensePlate: string;
  vehicleDescription: string | null;
  vehicleFeatures: VehicleFeature[];
};